+++ Assets/_Project/Scenes/WaterDemo.unity
@@ -130,6 +130,7 @@ GameObject:
   - component: {fileID: 590004144}
   - component: {fileID: 590004143}
   - component: {fileID: 590004142}
+  - component: {fileID: 590004145}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -211,6 +212,50 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &590004145
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 590004141}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a79441f348de89743a2939f4d699eac1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_RenderShadows: 1
+  m_RequiresDepthTextureOption: 2
+  m_RequiresOpaqueTextureOption: 2
+  m_CameraType: 0
+  m_Cameras: []
+  m_RendererIndex: -1
+  m_VolumeLayerMask:
+    serializedVersion: 2
+    m_Bits: 1
+  m_VolumeTrigger: {fileID: 0}
+  m_VolumeFrameworkUpdateModeOption: 2
+  m_RenderPostProcessing: 0
+  m_Antialiasing: 0
+  m_AntialiasingQuality: 2
+  m_StopNaN: 0
+  m_Dithering: 0
+  m_ClearDepth: 1
+  m_AllowXRRendering: 1
+  m_AllowHDROutput: 1
+  m_UseScreenCoordOverride: 0
+  m_ScreenSizeOverride: {x: 0, y: 0, z: 0, w: 0}
+  m_ScreenCoordScaleBias: {x: 0, y: 0, z: 0, w: 0}
+  m_RequiresDepthTexture: 0
+  m_RequiresColorTexture: 0
+  m_Version: 2
+  m_TaaSettings:
+    m_Quality: 3
+    m_FrameInfluence: 0.1
+    m_JitterScale: 1
+    m_MipBias: 0
+    m_VarianceClampScale: 0.9
+    m_ContrastAdaptiveSharpening: 0
 --- !u!1 &640499448
 GameObject:
   m_ObjectHideFlags: 0
@@ -1492,10 +1537,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 808315018}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: bb5adb15a9e854d4d8507d03754e8eb5, type: 3}
+  m_Script: {fileID: 11500000, guid: aae54a61a4c991748a87b4fd888ecb92, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  newTerrainData: {fileID: 11400000, guid: b16608c468fe9ae409bc3bc958142b93, type: 2}
+  terrainData: {fileID: 11400000, guid: d1f5a0da272866045a3ac0a1f86c7b56, type: 2}
   sourceTilemap: {fileID: 640499451}
   loadOnStart: 0
   elevationScale: 1
@@ -1525,11 +1570,13 @@ MonoBehaviour:
   m_GameObject: {fileID: 1982371192}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a102678634fce324792318472ce0808d, type: 3}
+  m_Script: {fileID: 11500000, guid: f2f1f82172468b84facb01d20fc380b3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   simulationManager: {fileID: 698237306}
   tilemap: {fileID: 1994088320}
+  terrainDataSource: {fileID: 11400000, guid: d1f5a0da272866045a3ac0a1f86c7b56, type: 2}
+  z_Booster: 0
   transparentTile: {fileID: 11400000, guid: 3d987b2eda58bb4479cde054d90b5b58, type: 2}
   lightTile: {fileID: 11400000, guid: 7bd16e3ea8c66ef4aa8f2ef51fbe8468, type: 2}
   mediumTile: {fileID: 11400000, guid: c761ee03f4ebb354dbc8a773533eb27b, type: 2}
@@ -1577,7 +1624,7 @@ Transform:
   m_GameObject: {fileID: 1994088317}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0.5, y: -1.75, z: 0}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
