--- Assets/_Project/Scripts/Water System/TileTypeSystem/FloodTilemapRenderer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a102678634fce324792318472ce0808d
